Revert "introduced render targets"
This reverts commit e55174becefd21cfff401b3b80c920700790a4aa.

+++ examples/04 3D/src/Example3D.h
@@ -97,27 +97,27 @@ private:
 
     /*! id to transform node used for manipulating the heading of all objects in scene
     */
-    uint64 _allObjectsHeading = iNode::INVALID_NODE_ID;
+    uint32 _allObjectsHeading = iNode::INVALID_NODE_ID;
 
     /*! id to transform node used for manipulating the pitch of all objects in scene
     */
-    uint64 _allObjectsPitch = iNode::INVALID_NODE_ID;
+    uint32 _allObjectsPitch = iNode::INVALID_NODE_ID;
 
     /*! id to transform node used for manipulating the heading of the camera
     */
-    uint64 _cameraHeading = iNode::INVALID_NODE_ID;
+    uint32 _cameraHeading = iNode::INVALID_NODE_ID;
 
     /*! id to transform node used for manipulating the pitch of the camera
     */
-    uint64 _cameraPitch = iNode::INVALID_NODE_ID;
+    uint32 _cameraPitch = iNode::INVALID_NODE_ID;
 
     /*! id to transform node used for manipulating the distance of the camera to the origin
     */
-    uint64 _cameraTranslation = iNode::INVALID_NODE_ID;
+    uint32 _cameraTranslation = iNode::INVALID_NODE_ID;
 
     /*! id to transform node used for manipulating the direction of the light relative to the origin
     */
-    uint64 _directionalLightRotate = iNode::INVALID_NODE_ID;
+    uint32 _directionalLightRotate = iNode::INVALID_NODE_ID;
 
     /*! timer handle to control the movement of the light source over time
     */
@@ -125,31 +125,31 @@ private:
 
     /*! id to crate model node
     */
-    uint64 _crateModel = iNode::INVALID_NODE_ID;
+    uint32 _crateModel = iNode::INVALID_NODE_ID;
 
     /*! id to cat model node
     */
-    uint64 _catModel = iNode::INVALID_NODE_ID;
+    uint32 _catModel = iNode::INVALID_NODE_ID;
 
     /*! id to teapot model node
     */
-    uint64 _teapotModel = iNode::INVALID_NODE_ID;
+    uint32 _teapotModel = iNode::INVALID_NODE_ID;
 
     /*! id to switch node
     */
-    uint64 _switchNode = iNode::INVALID_NODE_ID;
+    uint32 _switchNode = iNode::INVALID_NODE_ID;
 
     /*! index of switch position
     */
-    uint64 _activeNode = 0;
+    int32 _activeNode = 0;
 
     /*! id to lod trigger node
     */
-    uint64 _LODTrigger = iNode::INVALID_NODE_ID;
+    uint32 _LODTrigger = iNode::INVALID_NODE_ID;
 
     /*! id to lod switch node
     */
-    uint64 _LODSwitch = iNode::INVALID_NODE_ID;
+    uint32 _LODSwitch = iNode::INVALID_NODE_ID;
 
     /*! material definition for the sky box
     */
